Start refactoring schema info
