+++ TODO.txt
@@ -1,5 +1,3 @@
-* Clean up schema retrieval, SchemaInfo class population to prep for auto --skipPkStore feature
-
 * Automatic `--skipPkStore` option
   - Including case where a unique index means we might _not_ have to store PKs for a table
   - Include config option to print out which tables' PKs will be stored
