+++ src/main/scala/trw/dbsubsetter/db/SchemaInfoRetrieval.scala
@@ -1,136 +1,45 @@
 package trw.dbsubsetter.db
 
-import java.sql.{DriverManager, JDBCType}
 import java.util.NoSuchElementException
 
 import trw.dbsubsetter.config.Config
 
-import scala.collection.mutable.ArrayBuffer
-
 object SchemaInfoRetrieval {
   def getSchemaInfo(config: Config): SchemaInfo = {
-    val conn = DriverManager.getConnection(config.originDbConnectionString)
-    conn.setReadOnly(true)
-    val isMysql = conn.getMetaData.getDatabaseProductName == "MySQL"
-    val catalog = conn.getCatalog
-    val ddl = conn.getMetaData
+    val DbMetadataQueryResult(tables, columns, pks, fks, dbVendor) = DbMetadataQueries.queryDb(config)
 
-    val tablesQueryResult = ArrayBuffer.empty[TableQueryRow]
-    val columnsQueryResult = ArrayBuffer.empty[ColumnQueryRow]
-    val primaryKeysQueryResult = ArrayBuffer.empty[PrimaryKeyQueryRow]
-    val foreignKeysQueryResult = ArrayBuffer.empty[ForeignKeyQueryRow]
-
-    config.schemas.foreach { schema =>
-      // Args: catalog, schemaPattern, tableNamePattern, types
-      val tablesJdbcResultSet = {
-        if (isMysql) ddl.getTables(schema, null, "%", Array("TABLE"))
-        else ddl.getTables(catalog, schema, "%", Array("TABLE"))
-      }
-      while (tablesJdbcResultSet.next()) {
-        val table = tablesJdbcResultSet.getString("TABLE_NAME")
-        if (!config.excludeTables.contains((schema, table))) {
-          tablesQueryResult += TableQueryRow(schema, table)
-        }
-      }
-    }
-
-    tablesQueryResult.foreach { table =>
-      val colsJdbcResultSet = {
-        if (isMysql) ddl.getColumns(table.schema, null, table.name, "%")
-        else ddl.getColumns(catalog, table.schema, table.name, "%")
-      }
-      while (colsJdbcResultSet.next()) {
-        val columnName = colsJdbcResultSet.getString("COLUMN_NAME")
-        val jdbcType = JDBCType.valueOf(colsJdbcResultSet.getInt("DATA_TYPE"))
-        val typeName = colsJdbcResultSet.getString("TYPE_NAME")
-        val isSqlServerAutoIncrement = conn.isMsSqlServer && colsJdbcResultSet.getString("IS_AUTOINCREMENT") == "YES"
-
-        if (!config.excludeColumns((table.schema, table.name)).contains(columnName)) {
-          columnsQueryResult += ColumnQueryRow(table.schema, table.name, columnName, jdbcType, typeName, isSqlServerAutoIncrement)
-        }
-      }
-    }
-
-    // Args: catalog, schema, table
-    tablesQueryResult.foreach { table =>
-      val pksJdbcResultSet = {
-        if (isMysql) ddl.getPrimaryKeys(table.schema, null, table.name)
-        else ddl.getPrimaryKeys(catalog, table.schema, table.name)
-      }
-      while (pksJdbcResultSet.next()) {
-        primaryKeysQueryResult += PrimaryKeyQueryRow(
-          table.schema,
-          table.name,
-          pksJdbcResultSet.getString("COLUMN_NAME")
-        )
-      }
-    }
-
-    tablesQueryResult.foreach { table =>
-      // Args: catalog, schema, table
-      val foreignKeysJdbcResultSet = {
-        if (isMysql) ddl.getImportedKeys(table.schema, null, table.name)
-        else ddl.getImportedKeys(catalog, table.schema, table.name)
-      }
-      while (foreignKeysJdbcResultSet.next()) {
-        val fromSchema = if (isMysql) foreignKeysJdbcResultSet.getString("FKTABLE_CAT") else foreignKeysJdbcResultSet.getString("FKTABLE_SCHEM")
-        val fromTable = foreignKeysJdbcResultSet.getString("FKTABLE_NAME")
-        val fromColumn = foreignKeysJdbcResultSet.getString("FKCOLUMN_NAME")
-
-        val toSchema = if (isMysql) foreignKeysJdbcResultSet.getString("PKTABLE_CAT") else foreignKeysJdbcResultSet.getString("PKTABLE_SCHEM")
-        val toTable = foreignKeysJdbcResultSet.getString("PKTABLE_NAME")
-        val toColumn = foreignKeysJdbcResultSet.getString("PKCOLUMN_NAME")
-
-        if (!config.excludeTables.contains((fromSchema, fromTable)) && !config.excludeTables.contains((toSchema, toTable))) {
-          foreignKeysQueryResult += ForeignKeyQueryRow(fromSchema, fromTable, fromColumn, toSchema, toTable, toColumn)
-        }
-      }
-    }
-
-    val dbVendor = conn.dbVendor
-    conn.close()
-
-    config.cmdLinePrimaryKeys.foreach { clpk =>
-      clpk.columns.foreach(c => primaryKeysQueryResult += PrimaryKeyQueryRow(clpk.schema, clpk.table, c))
-    }
-
-    val tablesByName = tablesQueryResult.map { t =>
-      val hasSqlServerAutoincrement = columnsQueryResult.exists(c => c.schema == t.schema && c.table == t.name && c.isSqlServerAutoincrement)
+    val tablesByName = tables.map { t =>
+      val hasSqlServerAutoincrement = columns.exists(c => c.schema == t.schema && c.table == t.name && c.isSqlServerAutoincrement)
       val storePks = !config.skipPkStore.contains((t.schema, t.name))
       (t.schema, t.name) -> Table(t.schema, t.name, hasSqlServerAutoincrement, storePks)
     }.toMap
 
     val colsByTableAndName: Map[Table, Map[ColumnName, Column]] = {
-      columnsQueryResult
+      columns
         .groupBy(c => tablesByName(c.schema, c.table))
         .map { case (table, partialColumns) =>
           table -> partialColumns.zipWithIndex.map { case (pc, i) => pc.name -> Column(table, pc.name, i, pc.jdbcType, pc.typeName) }.toMap
         }
     }
+
     val colByTableOrdered: Map[Table, Vector[Column]] = {
       colsByTableAndName.map { case (table, map) => table -> map.values.toVector.sortBy(_.ordinalPosition) }
     }
 
-    val pkColumnOrdinalsByTable: Map[Table, Vector[Int]] = {
-      primaryKeysQueryResult
+    val pkOrdinalsByTable: Map[Table, Vector[Int]] = {
+      pks
         .groupBy(pk => tablesByName(pk.schema, pk.table))
         .map { case (table, partialPks) =>
-          table -> partialPks.map(ppk => colsByTableAndName(table)(ppk.column)).toVector.map(_.ordinalPosition).sorted
+          table -> partialPks.map(ppk => colsByTableAndName(table)(ppk.column)).map(_.ordinalPosition).sorted
         }
     }
 
     val foreignKeysOrdered: Array[ForeignKey] = {
-      val userSuppliedPartialFks: Seq[ForeignKeyQueryRow] = config.cmdLineForeignKeys.flatMap { cfk =>
-        cfk.fromColumns.zip(cfk.toColumns).map { case (fromCol, toCol) =>
-          ForeignKeyQueryRow(cfk.fromSchema, cfk.fromTable, fromCol, cfk.toSchema, cfk.toTable, toCol)
-        }
-      }
-      val allPartialFKs = userSuppliedPartialFks ++ foreignKeysQueryResult
-
-      val allFksUnordered = allPartialFKs.groupBy(fkm => (fkm.fromSchema, fkm.fromTable, fkm.toSchema, fkm.toTable))
-        .zipWithIndex.map { case (((fromSchemaName, fromTableName, toSchemaName, toTableName), partialForeignKeys), i) =>
+      val fksUnordered = fks
+        .groupBy(fkm => (fkm.fromSchema, fkm.fromTable, fkm.toSchema, fkm.toTable))
+        .map { case ((fromSchemaName, fromTableName, toSchemaName, toTableName), partialForeignKeys) =>
           val fromTable = tablesByName(fromSchemaName, fromTableName)
-          val fromCols = partialForeignKeys.map { pfk => colsByTableAndName(fromTable)(pfk.fromColumn) }.toVector
+          val fromCols = partialForeignKeys.map { pfk => colsByTableAndName(fromTable)(pfk.fromColumn) }
           val toTable = tablesByName(toSchemaName, toTableName)
           // MySQL schema introspection has a bug where they don't properly capitalize column names of
           // the `pointedTo` side of foreign keys.
@@ -150,20 +59,20 @@ object SchemaInfoRetrieval {
             try {
               colsByTableAndName(toTable)(pfk.toColumn)
             } catch {
-              case _: NoSuchElementException if isMysql => mysqlWorkaround(toTable)(pfk.toColumn)
+              case _: NoSuchElementException if dbVendor == DbVendor.MySQL => mysqlWorkaround(toTable)(pfk.toColumn)
               case e: Throwable => throw e
             }
-          }.toVector
+          }
 
           val pointsToPk = {
-            val pkOpt = pkColumnOrdinalsByTable.get(toTable)
+            val pkOpt = pkOrdinalsByTable.get(toTable)
             pkOpt.fold(false)(pkColOrdinals => pkColOrdinals == toCols.map(_.ordinalPosition))
           }
 
         ForeignKey(fromCols, toCols, pointsToPk, 0)
       }
 
-      allFksUnordered.toArray.zipWithIndex.map { case (fk, idx) => fk.copy(i = idx.toShort) }
+      fksUnordered.toArray.zipWithIndex.map { case (fk, idx) => fk.copy(i = idx.toShort) }
     }
 
     val fksFromTable: Map[Table, Vector[ForeignKey]] = {
@@ -177,33 +86,11 @@ object SchemaInfoRetrieval {
     SchemaInfo(
       tablesByName,
       colByTableOrdered,
-      pkColumnOrdinalsByTable,
+      pkOrdinalsByTable,
       foreignKeysOrdered,
       fksFromTable,
       fksToTable,
       dbVendor
     )
   }
-
-  private[this] case class TableQueryRow(schema: SchemaName,
-                                         name: TableName)
-
-  private[this] case class ColumnQueryRow(schema: SchemaName,
-                                          table: TableName,
-                                          name: ColumnName,
-                                          jdbcType: JDBCType,
-                                          typeName: String,
-                                          isSqlServerAutoincrement: Boolean)
-
-  private[this] case class PrimaryKeyQueryRow(schema: SchemaName,
-                                              table: TableName,
-                                              column: ColumnName)
-
-  private[this] case class ForeignKeyQueryRow(fromSchema: SchemaName,
-                                              fromTable: TableName,
-                                              fromColumn: ColumnName,
-                                              toSchema: SchemaName,
-                                              toTable: TableName,
-                                              toColumn: ColumnName)
-
 }
